bm-runner/bm_utils.py: place construct_bm_name here
